fat arrow expression added, syntactic sugar to define a two item tuple, read more below:
since a lot of the two item tuples are used to tag data I decided to add syntactic sugar to
make the code look better.

also since a way of building dict like structures is through a list of two item tuples where
the first item is the key and the second is the value, some syntactic sugar for that case wouldn't be bad right?

some examples:

>>> Nums = [one => 1, two => 2, fourtytwo => 42]
[{one,1},{two,2},{fourtytwo,42}]
>>> lists.keyfind(two, 1, Nums)
{two,2}
>>> lists.keymember(two, 1, Nums)
true
>>> lists.keymember(nine, 1, Nums)
false
>>> lists.keysearch(two, 1, Nums)
{value,{two,2}}

as you can see, the list module has a lot of functions to work on a data structure like this, and it looks
nice too ;)

this will also make DSL languages nicer in efene